Ensure the viewpoint selector is displayed above each element to prevent it from becoming unselectable. #192
Fixes #192
